fix: sync save buttons with the visible 'Saving…' indicator state
Revert the local isManualSaving flag. Move useDisplayStatus into the
SaveStatusProvider and expose displayStatus on the values context, so
the indicator and the Save Changes / Save now buttons all read the
same debounced view of the save state.

Effect: the buttons disable + show "Saving…" exactly while the blue
"Saving…" pill is on screen — including the 600ms minimum hold — and
revert as soon as it transitions to green/idle. Background autosaves
still flicker the buttons (intentional per user feedback) but the
flicker now lines up perfectly with the indicator.
